1582 Falla en recibo, no permite deshabilitar checks  - 1575 El reloj "Llevas cargando"no coicide con la suma de las balanzas - 1487 permitir decimales en cantidad de plano de carga por bodegas

+++ Molinos.Scato.Database/dbo/Tables/CargaComercialHistorico.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[CargaComercialHistorico] (
     [Id]                              INT             IDENTITY (1, 1) NOT NULL,
-    [Cantidad]					 INT             NOT NULL,
+    [Cantidad]					 DECIMAL(18, 3)             NOT NULL,
     [Exportador_Id]                       INT             NOT NULL,
     [MaterialPuerto_Id]                       INT             NOT NULL,
     [PlanoDeCargaHistorico_Id]            INT                      NOT NULL,
